pyproject.toml
[tool.poetry]
name = "iamwatchingyou"
version = "0.1.0"
description = ""
authors = ["GrandTiger1729 <80200287+GrandTiger1729@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.5.5"
requests = "^2.27.1"
numpy = "^1.22.4"
imutils = "^0.5.4"
python-dotenv = "^0.20.0"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
